Added history support to prompt widget.
Added three new classes, History, PeristentHistory and TransientHistory.
History is the base class that inherits from list, and handles common
functionality, i.e. handling key presses. PersistentHistory loads and
stores itself in $XDG_DATA_HOME/qtile/<prompt_name>.hist, and
TransientHistory has no storage.

Also changed is the addition of the `userentered` member of the Prompt
class. This is the index of the last character that the user entered.
This allows us to distinguish between data in the buffer that the user
has accepted, and data that is being shown to the user, by the History
or Completion classes.